Extract FileAttachmentsExample to app/examples like other examples
Move the FileAttachmentsExample (added by PR #21) from page.tsx into
its own file with an exported code string, wrap it in ExampleShowcase
with Preview/Code tabs, and use SectionHeading for consistency.
Also ignore .next/ in ESLint config to fix lint pre-commit hook.
