+++ app/examples/index.ts
@@ -7,6 +7,7 @@ export { AsyncSearchExample, asyncSearchCode } from './async-search'
 export { MarkdownExample, markdownCode } from './markdown'
 export { CopyPasteExample, copyPasteCode } from './copy-paste'
 export { ImageAttachmentsExample, imageAttachmentsCode } from './image-attachments'
+export { FileAttachmentsExample, fileAttachmentsCode } from './file-attachments'
 export { ActionBarFullExample, actionBarFullCode } from './action-bar-full'
 export { ActionBarMinimalExample, actionBarMinimalCode } from './action-bar-minimal'
 export { ActionBarDisabledExample, actionBarDisabledCode } from './action-bar-disabled'
